Properly setup/teardown a roster.test.db during test runs.
I had to introduce a RUNNING_TESTS boolean into the db.py code
that I'm not terribly proud of, but I'm not sure what a better
solution is.

db.py also needs to use a NullPool when running tests because
of http://stackoverflow.com/a/6135064.

Also, I changed roster.unique_name_id to have a session object
passed into it, as keeping the lifecycle of a session separate and
external from functions that use it is apparently a best practice with
SQLAlchemy.